requirements.txt
enum34==1.1.2
msgpack-python==0.4.7
ngram==3.3.0
puzpy==0.2.1
py==1.4.31
pytest==2.8.7
wheel==0.24.0
